refactor: Split code into various plugins
